Solve numpy-arange-linspace

+++ study-plans/numpy-basics/numpy-arange-linspace/numpy-arange-linspace.py
@@ -6,4 +6,4 @@ def create_sequence(start, stop, param, kind):
     """
     if kind =="arange":
         return np.arange(start, stop, param, dtype= np.float64)
-    return np.linspace(start, stop, param,dtype= np.float64)
+    return np.linspace(start, stop, int(param))
